requirements.txt
faiss-cpu==1.7.4
Flask==2.3.3
Flask-Cors==4.0.0
huggingface-hub==0.17.1
numpy==1.25.2
openai==1.14.0
langchain==0.1.12
pandas==2.1.0
tqdm==4.66.1
typing_extensions==4.7.1
python-dotenv==1.0.0
loguru==0.7.2
pytest>=7.2.1
sentence_transformers==2.2.2